Update README.md
This project involves building a relational database that replicates the structure of Yelp’s production database, designed specifically to support experimentation with advanced machine learning tools. The database contains a curated subset of Yelp data to enable faster prototyping and model development. Along with the database, comprehensive documentation—including a detailed data dictionary and an Entity-Relationship Diagram (ERD)—was created to help the data science team easily navigate and utilize the data. The project also includes example query analyses and visualizations (graphs and charts) to assist in exploring the data and generating insights.